geobackend/requirements.txt
asgiref==3.7.2
Django==4.0.2
django-colorfield==0.10.1
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.13.1
Pillow==10.0.1
psycopg2-binary==2.9.3
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
